# mm/mm_heap : algorithm enhancement to reduce temporal cost of malloc()