--- a/apps/examples/heap_performance_test/Kconfig
+++ b/apps/examples/heap_performance_test/Kconfig
@@ -1,0 +1,14 @@
+#
+# For a description of the syntax of this configuration file,
+# see kconfig-language at https://www.kernel.org/doc/Documentation/kbuild/kconfig-language.txt
+#
+
+config EXAMPLES_HEAP_PERFORMANCE_TEST
+	bool "Heap performance test"
+	default n
+	---help---
+		Measure the elapsed time while simply repeating memory allocation and release.
+
+config USER_ENTRYPOINT
+	string
+	default "heaptest_main" if ENTRY_HEAP_PERFORMANCE_TEST