--- a/apps/examples/heap_performance_test/Kconfig_ENTRY
+++ b/apps/examples/heap_performance_test/Kconfig_ENTRY
@@ -1,0 +1,3 @@
+config ENTRY_HEAP_PERFORMANCE_TEST
+	bool "Heap performance test"
+	depends on EXAMPLES_HEAP_PERFORMANCE_TEST