--- a/apps/examples/heap_performance_test/README.txt
+++ b/apps/examples/heap_performance_test/README.txt
@@ -1,0 +1,7 @@
+examples/heap_performance_test
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+  This is an example to measure the elapsed time while simply repeating memory allocation and release.
+  
+  Configs (see the details on Kconfig):
+  * CONFIG_EXAMPLES_HEAP_PERFORMANCE_TEST