--- a/os/mm/mm_heap/mm_addfreechunk.c
+++ b/os/mm/mm_heap/mm_addfreechunk.c
@@ -88,9 +88,9 @@ void mm_addfreechunk(FAR struct mm_heap_s *heap, FAR struct mm_freenode_s *node)
 
 	int ndx = mm_size2ndx(node->size);
 
-	/* Now put the new node int the next */
+	/* Now put the new free node in a descending order */
 
-	for (prev = &heap->mm_nodelist[ndx], next = heap->mm_nodelist[ndx].flink; next && next->size && next->size < node->size; prev = next, next = next->flink) ;
+	for (prev = &heap->mm_nodelist[ndx], next = prev->flink; next && next->size > node->size; prev = next, next = next->flink) ;
 
 	/* Does it go in mid next or at the end? */