Current algorithm:
Free nodes are managed by n double-linked lists called 'mm_nodelist'.
The head of each list is not a real free node but discrimiates where each list starts.
i-th list includes free nodes whose size ranges [2^(i+4) 2^(i+5)[.
When memory allocation is requested, the current procedure is as follows:
1.compute 'ndx' with size
2.search an available node from 'mm_nodelist[ndx]
3.remove the node from the list.

Here are two problems. At the second step, we may not find an available node from ndx-th list.
Let's see an example.
The table depicted below shows the initial status of the heap memory.
When 32 bytes allocation is requested, from the step 1 the index is 1.
However, mm_nodelist[1] doesn't have any free nodes.
To handle this situation, each head is doubly-linked with adjacent head nodes at initialization.
So, searching can continue and go through mm_nodelist[2] up to mm_nodelist[15].
Finally, one available free node is found at mm_nodelist[15].
This exhaustive search just increases the temporal cost of malloc() operations.

Nodelist[0] (less than 32) : num 0, size 0 [Bytes]
Nodelist[1] (less than 64) : num 0, size 0 [Bytes]
Nodelist[2] (less than 128) : num 0, size 0 [Bytes]
Nodelist[3] (less than 256) : num 0, size 0 [Bytes]
Nodelist[4] (less than 512) : num 0, size 0 [Bytes]
Nodelist[5] (less than 1024) : num 0, size 0 [Bytes]
Nodelist[6] (less than 2048) : num 0, size 0 [Bytes]
Nodelist[7] (less than 4096) : num 0, size 0 [Bytes]
Nodelist[8] (less than 8192) : num 0, size 0 [Bytes]
Nodelist[9] (less than 16384) : num 0, size 0 [Bytes]
Nodelist[10] (less than 32768) : num 0, size 0 [Bytes]
Nodelist[11] (less than 65536) : num 0, size 0 [Bytes]
Nodelist[12] (less than 131072) : num 0, size 0 [Bytes]
Nodelist[13] (less than 262144) : num 0, size 0 [Bytes]
Nodelist[14] (less than 524288) : num 0, size 0 [Bytes]
Nodelist[15] (less than 1048576) : num 1, size 897600 [Bytes]
Nodelist[16] (less than 2097152) : num 0, size 0 [Bytes]
Nodelist[17] (less than 4194304) : num 0, size 0 [Bytes]
Nodelist[18] (less than 8388608) : num 0, size 0 [Bytes]

The second problem is more severe.
Let's see another example.
When 48 bytes allocation is requested, from the step 1 the index is 1.
mm_nodelist[1] already has 209 free nodes but all free nodes are 32 bytes-sized.
After exhaustive searching through mm_nodelist[1],
finally, one available free node, which is larger than or equal to 48 bytes,
is found at mm_nodelist[15].

Nodelist[0] (less than 32) : num 0, size 0 [Bytes]
Nodelist[1] (less than 64) : num 209, size 6688 [Bytes]
Nodelist[2] (less than 128) : num 0, size 0 [Bytes]
Nodelist[3] (less than 256) : num 0, size 0 [Bytes]
...
Nodelist[15] (less than 1048576) : num 1, size 890912 [Bytes]

New algorithm:
We're going to simply resolve this problem.
Here are our approaches:
1. To tackle the first problem, delink Nodelist[n] from Nodelist[n-1] and Nodelist[n+1].
2. To handle the second problem, each list manages free nodes not in an ascending order but in a descending order in terms of size.
That is, the first free node is the largest one in the list.
This prevents exhaustive search through the whole free nodes in the list.
Now, just look up the first node and see whether the list has an available free node.
3. If free nodes are sorted in a descending order, it would be better to use ]32, 64] instead of range [32, 64[.
Please, see the table composed by this new rule.

Nodelist[0] (less than or equal to 32) : num 0, size 0 [Bytes]
Nodelist[1] (less than or equal to 64) : num 0, size 0 [Bytes]
Nodelist[2] (less than or equal to 128) : num 0, size 0 [Bytes]

Performance (Temporal cost):
We do an experiment where m bytes memory is allocated 100 times continuously and then all allocated memory blocks are released.
We measure the elapsed time for doing this experiment 1000 times continuously on ARTIK 053.
Compared to the previous one, the proposed algorithm shows much stable temporal performance.
That is, one pair of malloc() and free() takes almost the same time regardless of sizes.
With heavy load and size of 32, the previous one shows that the time per one pair of malloc() and free() is 15.17 us which is almost larger three times compared to other sizes.
In the same case, the proposed algorithm has up to 68.45 percent reduction in temporal cost.

Time per malloc/free [us] with light load
size [bytes] | 16   | 32    | 64    | 128  | 256  | 512  | 1024 | 2048 | 4096 | 8192
before [us] | 3.79 | 3.79 | 3.99 | 4.1   | 4.19 | 4.3   | 4.29   | 4.39  | 4.29  | 4.29
after    [us] | 3.59 | 3.69 | 3.69 | 3.89 | 4.0   | 3.99 | 4.0     | 3.99  | 4.09  | 4.0

Time per malloc/free [us] with heavy load
size [bytes] | 16   | 32      | 64    | 128  | 256  | 512 | 1024 | 2048 | 4096
before [us] | 5.08 | 15.17 | 6.58 | 5.09 | 7.09 | 5.98 | 5.29  | 4.59  | 4.79
after    [us] | 3.09 | 4.77   | 4.19 | 3.7   | 3.89 | 4.09 | 4.09  | 5.08  | 4.7

On Mediatek, the almost same results are shown.
The proposed algorithm has up to 58.12 percent reduction in temporal cost in case of heavy load at size of 32.

Time per malloc/free [us] with heavy load
size [bytes] | 16   | 32      | 64      | 128  | 256  | 512  | 1024 | 2048  | 4096  | 8192
before [us] | 7.25 | 26.84 | 10.91 | 8.49 | 8.58 | 8.52  | 8.85  | 12.03 | 12.73 | 9.22
after    [us] | 7.32 | 11.24 | 9.39   | 8.80 | 8.74 | 8.88  | 9.32  | 12.02 | 11.23 | 9.86