@seinfra 2019-05-11
Target : [9dfbbd78b2c037c34591c84432d23729b300677b] - Code Rule Check (C++) OK.