@sunghan-chang 2019-06-05
@davidfather Could you remove latest commit, `Merge branch 'master' into mm_heap`?
To resolve conflict, you should use `git rebase` instead of `git pull` or `git merge`.
```
$ git reset --hard HEAD^
$ git fetch upsteam
$ git rebase upstream/master
```