@seinfra 2019-06-05
Target : [602388e08094bb16f7d14b2fbaa0fe4226652445] - Code Rule Check (C++) OK.